requirements.txt
altair==5.4.1
annotated-types==0.7.0
attrs==24.2.0
blinker==1.8.2
blis==1.0.1
cachetools==5.5.0
catalogue==2.0.10
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
cloudpathlib==0.20.0
confection==0.1.5
cymem==2.0.8
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
filelock==3.16.1
fsspec==2024.10.0
gitdb==4.0.11
GitPython==3.1.43
huggingface-hub==0.26.2
idna==3.10
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
langcodes==3.4.1
language_data==1.2.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
murmurhash==1.0.10
narwhals==1.13.2
nltk==3.9.1
numpy==2.0.2
opencv-python==4.10.0.84
packaging==24.1
pandas==2.2.3
pillow==11.0.0
preshed==3.0.9
protobuf==5.28.3
pyarrow==18.0.0
pydantic==2.9.2
pydantic_core==2.23.4
pydeck==0.9.1
Pygments==2.18.0
python-dateutil==2.9.0.post0
pytube==15.0.0
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rpds-py==0.21.0
safetensors==0.4.5
setuptools==75.3.0
shellingham==1.5.4
six==1.16.0
smart-open==7.0.5
smmap==5.0.1
spacy==3.8.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SpeechRecognition==3.11.0
srsly==2.4.8
streamlit==1.40.0
tenacity==9.0.0
textblob==0.18.0.post0
thinc==8.3.2
tokenizers==0.20.3
toml==0.10.2
tornado==6.4.1
tqdm==4.67.0
transformers==4.46.2
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wasabi==1.1.3
weasel==0.4.1
wrapt==1.16.0
